+++ changelog.md
@@ -1,5 +1,9 @@
 <h3>Changelog</h3>
 
+## v138 - January 9, 2024
+
+* Bind toggle to `keydown` instead of `keyup` so that pressing `Ctrl+W` to close a tab does not trigger the YTWP toggle when you then focus on a Youtube tab.
+
 ## v137 - November 1, 2023
 
 * Bind to `yt-page-data-fetched` and `yt-navigate-finish` to fix the back button not cleaning up the window view since `yt-navigate-start` does not always fire (Issue #72 and #76)